SearchRepositoryAdapterを作成
